Add files via upload
The goal of this capstone project is to allow us to explore Python 3 via a project to embed the learning from the lectures and exercises.
The Health Service Executive (HSE) will deploy low-cost, terminal-based kiosks that will allow the public access to basic information about their various premises. The kiosks will initially be developed on a, Debian GNU/Linux-based operating system installed on Raspberry Pi devices. These devices have the potential
to be employed as embedded devices for the mass rollout.